requirements.txt
Django==1.5.4
argparse==1.2.1
distribute==0.6.34
django-social-auth==0.7.28
google-api-python-client==1.2
httplib2==0.8
numpy==1.7.1
oauth2==1.5.211
oauthlib==0.6.0
python-openid==2.2.5
python-social-auth==0.1.13
requests==2.0.0
requests-oauthlib==0.4.0
six==1.4.1
wsgiref==0.1.2
